refactor(chat): extract shared disposition state logic into useDispositionState hook
Eliminates verbatim duplication of controlledDispositions state, the
controlToken effect, and dispositionState derived logic across HealingNode
and RollSummary. Both components now delegate to useDispositionState,
passing getters for their targetDisposition and target UUIDs.
