+++ src/view/chat/components/HealingNode.svelte
@@ -2,6 +2,7 @@
 	import { getContext } from 'svelte';
 	import prepareRollTooltip from '../../dataPreparationHelpers/rollTooltips/prepareRollTooltip.js';
 	import type { NimbleChatMessage, AppliedHealingRecord } from '../../../documents/chatMessage.js';
+	import { useDispositionState } from '../utils/useDispositionState.svelte.ts';
 
 	const messageDocument = getContext('messageDocument') as NimbleChatMessage;
 	const { actorType, permissions } = messageDocument.system as {
@@ -11,11 +12,6 @@
 	const { healingTypes } = CONFIG.NIMBLE;
 
 	let { node } = $props();
-
-	// Tracks the dispositions of currently canvas-controlled tokens, updated live via
-	// the controlToken hook so the button responds to token selection without requiring
-	// the user to explicitly add a target to the message first.
-	let controlledDispositions = $state<number[]>([]);
 
 	let healingType = $derived(node.healingType);
 	let roll = $derived(node.roll);
@@ -41,56 +37,10 @@
 		node.targetDisposition as 'friendly' | 'neutral' | 'hostile' | 'secret' | undefined,
 	);
 
-	// Returns 'recommended' when the effective targets match this effect's targetDisposition,
-	// 'discouraged' when all targets are clearly the opposite type (hostile targets for a
-	// heal-only effect), or 'neutral' otherwise (no targets, neutral tokens, mixed).
-	// Both buttons are always visible — only the emphasis changes.
-	let dispositionState = $derived.by((): 'recommended' | 'neutral' | 'discouraged' => {
-		if (!targetDisposition) return 'neutral';
-
-		const dispMap: Record<string, number> = {
-			friendly: CONST.TOKEN_DISPOSITIONS.FRIENDLY,
-			neutral: CONST.TOKEN_DISPOSITIONS.NEUTRAL,
-			hostile: CONST.TOKEN_DISPOSITIONS.HOSTILE,
-			secret: CONST.TOKEN_DISPOSITIONS.SECRET,
-		};
-		const required = dispMap[targetDisposition];
-
-		const allDispositions: number[] = [...controlledDispositions];
-		for (const uuid of systemData.targets ?? []) {
-			const doc = fromUuidSync(uuid) as TokenDocument | null;
-			const disp = (doc as { disposition?: number } | null)?.disposition;
-			if (typeof disp === 'number') allDispositions.push(disp);
-		}
-
-		if (allDispositions.length === 0) return 'neutral';
-		if (allDispositions.some((d) => d === required)) return 'recommended';
-
-		// Discouraged only when every target is clearly the opposing type
-		const oppositeDisp: Partial<Record<string, number>> = {
-			friendly: CONST.TOKEN_DISPOSITIONS.HOSTILE,
-			hostile: CONST.TOKEN_DISPOSITIONS.FRIENDLY,
-		};
-		const opposite = oppositeDisp[targetDisposition];
-		if (opposite !== undefined && allDispositions.every((d) => d === opposite))
-			return 'discouraged';
-
-		return 'neutral';
-	});
-
-	$effect(() => {
-		function syncControlled() {
-			controlledDispositions = (canvas?.tokens?.controlled ?? []).map((t) => {
-				const doc = t.document as TokenDocument | null;
-				return doc?.disposition ?? CONST.TOKEN_DISPOSITIONS.FRIENDLY;
-			});
-		}
-		syncControlled();
-		const hookId = Hooks.on('controlToken', syncControlled);
-		return () => {
-			Hooks.off('controlToken', hookId);
-		};
-	});
+	const { dispositionState } = useDispositionState(
+		() => targetDisposition,
+		() => systemData.targets ?? [],
+	);
 
 	// Localization
 	const localize = (key: string) => game.i18n.localize(`NIMBLE.chat.${key}`);
