+++ src/view/chat/components/RollSummary.svelte
@@ -4,6 +4,7 @@
 
 	import { getContext } from 'svelte';
 	import localize from '#utils/localize.ts';
+	import { useDispositionState } from '../utils/useDispositionState.svelte.ts';
 
 	const {
 		label,
@@ -19,9 +20,6 @@
 	const autoExpand = game.settings.get('nimble', 'autoExpandRolls');
 	const messageDocument = getContext<NimbleChatMessage | undefined>('messageDocument');
 	let expanded = $state(autoExpand);
-	// Tracks the dispositions of currently canvas-controlled tokens so the button responds
-	// to token selection in real time without requiring an explicit "Add Target" click.
-	let controlledDispositions = $state<number[]>([]);
 
 	let canApplyDamage = $derived.by(() => {
 		if (type !== 'damage' || !game.user?.isGM) return false;
@@ -36,58 +34,12 @@
 		canApplyDamage ? applyDamageLabel : localize('NIMBLE.chat.noDamageToApply'),
 	);
 
-	// Returns 'recommended' when the effective targets match this effect's targetDisposition,
-	// 'discouraged' when all targets are clearly the opposite type, or 'neutral' otherwise.
-	// Both buttons are always visible — only the emphasis changes.
-	let dispositionState = $derived.by((): 'recommended' | 'neutral' | 'discouraged' => {
-		if (!targetDisposition) return 'neutral';
-
-		const dispMap: Record<string, number> = {
-			friendly: CONST.TOKEN_DISPOSITIONS.FRIENDLY,
-			neutral: CONST.TOKEN_DISPOSITIONS.NEUTRAL,
-			hostile: CONST.TOKEN_DISPOSITIONS.HOSTILE,
-			secret: CONST.TOKEN_DISPOSITIONS.SECRET,
-		};
-		const required = dispMap[targetDisposition];
-
-		const messageTargets =
+	const { dispositionState } = useDispositionState(
+		() => targetDisposition,
+		() =>
 			(messageDocument?.reactive as unknown as { system?: { targets?: string[] } } | undefined)
-				?.system?.targets ?? [];
-		const allDispositions: number[] = [...controlledDispositions];
-		for (const uuid of messageTargets) {
-			const doc = fromUuidSync(uuid) as TokenDocument | null;
-			const disp = (doc as { disposition?: number } | null)?.disposition;
-			if (typeof disp === 'number') allDispositions.push(disp);
-		}
-
-		if (allDispositions.length === 0) return 'neutral';
-		if (allDispositions.some((d) => d === required)) return 'recommended';
-
-		// Discouraged only when every target is clearly the opposing type
-		const oppositeDisp: Partial<Record<string, number>> = {
-			friendly: CONST.TOKEN_DISPOSITIONS.HOSTILE,
-			hostile: CONST.TOKEN_DISPOSITIONS.FRIENDLY,
-		};
-		const opposite = oppositeDisp[targetDisposition];
-		if (opposite !== undefined && allDispositions.every((d) => d === opposite))
-			return 'discouraged';
-
-		return 'neutral';
-	});
-
-	$effect(() => {
-		function syncControlled() {
-			controlledDispositions = (canvas?.tokens?.controlled ?? []).map((t) => {
-				const doc = t.document as TokenDocument | null;
-				return doc?.disposition ?? CONST.TOKEN_DISPOSITIONS.FRIENDLY;
-			});
-		}
-		syncControlled();
-		const hookId = Hooks.on('controlToken', syncControlled);
-		return () => {
-			Hooks.off('controlToken', hookId);
-		};
-	});
+				?.system?.targets ?? [],
+	);
 </script>
 
 <div class="roll" class:roll--no-subheading={!subheading}>
